remove new_transfer_popup + separate RadioButton and BorderedContainer widgets into their own file
